Refactored CLI commands into separate script files
