+++ commands/index.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  generate:     require('./generate'),
+  package:      require('./package'),
+  run:          require('./run'),
+  visualstudio: require('./visualstudio')
+};
